edit files

+++ src/main/java/com/example/pokemon_tcg/services/implementations/PokemonServiceImpl.java
@@ -1,6 +1,6 @@
 package com.example.pokemon_tcg.services.implementations;
 
-
+import com.example.pokemon_tcg.constants.TypePokemon;
 import com.example.pokemon_tcg.models.Pokemon;
 import com.example.pokemon_tcg.repositories.PokemonRepository;
 import com.example.pokemon_tcg.services.IPokemonService;
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class PokemonServiceImpl implements IPokemonService {
@@ -21,14 +22,14 @@ public class PokemonServiceImpl implements IPokemonService {
     }
 
     @Override
-    public Pokemon getPokemonById(String uuid) {
-        return pokemonRepository.findById(uuid)
-                .orElseThrow(() -> new RuntimeException("Pokemon non trouvé"));
+    public List<Pokemon> getAllPokemon(TypePokemon type) {
+        return (type == null) ? pokemonRepository.findAll() : pokemonRepository.findAllByType(type);
     }
 
     @Override
-    public List<Pokemon> getPokemonByType(String type) {
-        return pokemonRepository.findByType(type);
+    public Pokemon getPokemonById(String uuid) {
+        return pokemonRepository.findById(uuid)
+                .orElseThrow(() -> new RuntimeException("Pokemon non trouvé"));
     }
 
     @Override
@@ -37,7 +38,12 @@ public class PokemonServiceImpl implements IPokemonService {
     }
 
     @Override
-    public void deletePokemon(String uuid) {
-        pokemonRepository.deleteById(uuid);
+    public boolean deletePokemon(String uuid) {
+        Optional<Pokemon> pokemon = pokemonRepository.findById(uuid);
+        if (pokemon.isPresent()) {
+            pokemonRepository.deleteById(uuid);
+            return true;
+        }
+        return false;
     }
 }
